requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.33.3
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
netifaces==0.10.6
numpy==2.2.3
ombott==2.4
openai==1.63.2
pandas==2.2.3
pluralize==20240519.3
portalocker==3.1.1
py4web==1.20250215.1
pycparser==2.22
pycryptodome==3.21.0
pydal==20250215.2
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.8.0
pytz==2025.1
pywin32==308
renoir==1.8.0
requests==2.32.3
rocket3==20241225.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
threadsafevariable==20230507.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
waitress==3.0.2
watchgod==0.8.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yatl==20230507.3
zope.event==5.0
zope.interface==7.2
